// source/api/models/Company.model.ts
import mongoose from "mongoose";

export interface CompanyData {
    title: string;
    categoryID: string | CompanyDocument;
    status: boolean;
    description: string;
    image: string;
}

export interface CompanyDocument extends CompanyData, mongoose.Document {
    createdAt: Date;
    updatedAt: Date;
}

export const companySchema = new mongoose.Schema(
    {
        // ID is automatically generated by mongoDB
        title: {
            type: String,
            required: true,
        },
        categoryID: {
            type: mongoose.Schema.Types.ObjectId,
            required: false,
            ref: "companyCategory",
        },
        description: {
            type: String,
            required: false,
        },
        image: {
            type: String,
            required: false,
        },
        status: {
            type: Boolean,
            required: false,
        },
    },
    {
        // createdAt & updatedAt is also automatically generated
        timestamps: true,
    }
);

const Company = mongoose.model<CompanyDocument>("company", companySchema);

export default Company;